+++ app/sorters/dorsale/alexandrie/attachments_sorter.rb
@@ -4,11 +4,11 @@ class Dorsale::Alexandrie::AttachmentsSorter < Agilibox::Sorter
     when :created_at, :updated_at
       {column => direction}
     when :name
-      %(LOWER(#{model.table_name}.#{column}) #{direction})
+      Arel.sql %(LOWER(#{model.table_name}.#{column}) #{direction})
     when :attachment_type_name
       @collection = @collection.joins(:attachment_type)
       table = Dorsale::Alexandrie::AttachmentType.table_name
-      %(LOWER(#{table}.name) #{direction})
+      Arel.sql %(LOWER(#{table}.name) #{direction})
     else
       {created_at: :desc, id: :desc}
     end